+++ packages/web/src/javascripts/Components/NoteView/NoteView.tsx
@@ -971,6 +971,7 @@ class NoteView extends PureComponent<NoteViewProps, State> {
                     navigationController={this.viewControllerManager.navigationController}
                     notesController={this.viewControllerManager.notesController}
                     noteTagsController={this.viewControllerManager.noteTagsController}
+                    historyModalController={this.viewControllerManager.historyModalController}
                     onClickPreprocessing={this.ensureNoteIsInsertedBeforeUIAction}
                   />
                 </div>
@@ -1002,7 +1003,6 @@ class NoteView extends PureComponent<NoteViewProps, State> {
                   onLoad={this.onEditorComponentLoad}
                   requestReload={this.editorComponentViewerRequestsReload}
                   application={this.application}
-                  viewControllerManager={this.viewControllerManager}
                 />
               </div>
             )}
@@ -1073,12 +1073,7 @@ class NoteView extends PureComponent<NoteViewProps, State> {
               {this.state.stackComponentViewers.map((viewer) => {
                 return (
                   <div className="component-view component-stack-item" key={viewer.identifier}>
-                    <ComponentView
-                      key={viewer.identifier}
-                      componentViewer={viewer}
-                      application={this.application}
-                      viewControllerManager={this.viewControllerManager}
-                    />
+                    <ComponentView key={viewer.identifier} componentViewer={viewer} application={this.application} />
                   </div>
                 )
               })}